+++ src/py/mat3ra/esse/models/materials_category/defects/two_dimensional/slab_grain_boundary/configuration.py
@@ -410,7 +410,7 @@ class SupercellMatrix2DSchemaItem(RootModel[List[int]]):
 
 class SlabConfigurationSchema(BaseModel):
     bulk: Optional[MaterialSchema] = Field(None, title="material schema")
-    miller_indices: Optional[List[float]] = Field([0, 0, 1], max_length=3, min_length=3, title="Miller Indices Schema")
+    miller_indices: Optional[List[int]] = Field([0, 0, 1], max_length=3, min_length=3, title="Miller Indices Schema")
     """
     Miller indices for crystallographic plane designation
     """
@@ -719,7 +719,7 @@ class MaterialSchema18(BaseModel):
 
 class SlabConfigurationSchema3(BaseModel):
     bulk: Optional[MaterialSchema18] = Field(None, title="material schema")
-    miller_indices: Optional[List[float]] = Field([0, 0, 1], max_length=3, min_length=3, title="Miller Indices Schema")
+    miller_indices: Optional[List[int]] = Field([0, 0, 1], max_length=3, min_length=3, title="Miller Indices Schema")
     """
     Miller indices for crystallographic plane designation
     """
@@ -1039,7 +1039,7 @@ class MaterialSchema19(BaseModel):
 
 class SlabConfigurationSchema4(BaseModel):
     bulk: Optional[MaterialSchema19] = Field(None, title="material schema")
-    miller_indices: Optional[List[float]] = Field([0, 0, 1], max_length=3, min_length=3, title="Miller Indices Schema")
+    miller_indices: Optional[List[int]] = Field([0, 0, 1], max_length=3, min_length=3, title="Miller Indices Schema")
     """
     Miller indices for crystallographic plane designation
     """
